refactor: workflow queue service
