+++ apps/api/src/app/events/services/workflow.queue.service.ts
@@ -1,13 +1,9 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { differenceInMilliseconds } from 'date-fns';
 import { Queue, Worker, QueueBaseOptions, JobsOptions, QueueScheduler } from 'bullmq';
-import { SendMessage } from '../usecases/send-message/send-message.usecase';
-import { SendMessageCommand } from '../usecases/send-message/send-message.command';
-import { QueueNextJob } from '../usecases/queue-next-job/queue-next-job.usecase';
-import { QueueNextJobCommand } from '../usecases/queue-next-job/queue-next-job.command';
 import { JobEntity, JobRepository, JobStatusEnum } from '@novu/dal';
-import { StepTypeEnum, DigestUnitEnum, DelayTypeEnum } from '@novu/shared';
-import { ApiException } from '../../shared/exceptions/api.exception';
+import { DigestUnitEnum } from '@novu/shared';
+import { RunJob } from '../usecases/run-job/run-job.usecase';
+import { RunJobCommand } from '../usecases/run-job/run-job.command';
 
 interface IJobEntityExtended extends JobEntity {
   presend?: boolean;
@@ -28,11 +24,9 @@ export class WorkflowQueueService {
   public readonly queue: Queue;
   public readonly worker: Worker;
   @Inject()
-  private sendMessage: SendMessage;
-  @Inject()
-  private queueNextJob: QueueNextJob;
-  @Inject()
   private jobRepository: JobRepository;
+  @Inject()
+  private runJob: RunJob;
   private readonly queueScheduler: QueueScheduler;
 
   constructor() {
@@ -64,35 +58,10 @@ export class WorkflowQueueService {
   }
 
   public async work(job: IJobEntityExtended) {
-    const canceled = await this.delayedEventIsCanceled(job);
-    if (canceled) {
-      return;
-    }
-
-    await this.jobRepository.updateStatus(job._id, JobStatusEnum.RUNNING);
-
-    await this.sendMessage.execute(
-      SendMessageCommand.create({
-        identifier: job.identifier,
-        payload: job.payload ? job.payload : {},
-        overrides: job.overrides ? job.overrides : {},
-        step: job.step,
-        transactionId: job.transactionId,
-        notificationId: job._notificationId,
-        environmentId: job._environmentId,
-        organizationId: job._organizationId,
-        userId: job._userId,
-        subscriberId: job._subscriberId,
+    await this.runJob.execute(
+      RunJobCommand.create({
         jobId: job._id,
-        events: job.digest.events,
-      })
-    );
-    if (job.presend === true) {
-      return;
-    }
-    await this.queueNextJob.execute(
-      QueueNextJobCommand.create({
-        parentId: job._id,
+        presend: job.presend,
         environmentId: job._environmentId,
         organizationId: job._organizationId,
         userId: job._userId,
@@ -100,31 +69,13 @@ export class WorkflowQueueService {
     );
   }
 
-  public async addJob(data: JobEntity | undefined, presend = false) {
-    if (!data) {
-      return;
-    }
+  public async addToQueue(id: string, data: any, delay?: number | undefined) {
     const options: JobsOptions = {
       removeOnComplete: true,
       removeOnFail: true,
+      delay,
     };
-
-    const digestAdded = await this.addDigestJob(data, options);
-    const delayAdded = await this.addDelayJob(data, options);
-
-    if (digestAdded || delayAdded) {
-      return;
-    }
-
-    await this.jobRepository.updateStatus(data._id, JobStatusEnum.QUEUED);
-    await this.queue.add(
-      data._id,
-      {
-        ...data,
-        presend,
-      },
-      options
-    );
+    await this.queue.add(id, data, options);
   }
 
   public static toMilliseconds(amount: number, unit: DigestUnitEnum): number {
@@ -140,100 +91,5 @@ export class WorkflowQueueService {
     }
 
     return delay;
-  }
-
-  private async addDigestJob(data: JobEntity, options: JobsOptions): Promise<boolean> {
-    const isValidDigestStep = data.type === StepTypeEnum.DIGEST && data.digest.amount && data.digest.unit;
-    if (!isValidDigestStep) {
-      return false;
-    }
-
-    const where: Partial<JobEntity> = {
-      status: JobStatusEnum.DELAYED,
-      type: StepTypeEnum.DIGEST,
-      _subscriberId: data._subscriberId,
-      _templateId: data._templateId,
-      _environmentId: data._environmentId,
-    };
-    const delayedDigest = await this.jobRepository.findOne(where);
-
-    if (delayedDigest) {
-      return true;
-    }
-
-    await this.jobRepository.updateStatus(data._id, JobStatusEnum.DELAYED);
-    const delay = WorkflowQueueService.toMilliseconds(data.digest.amount, data.digest.unit);
-    if (data.digest?.updateMode) {
-      const inApps = await this.jobRepository.findInAppsForDigest(data.transactionId, data._subscriberId);
-      for (const inApp of inApps) {
-        await this.addJob(inApp, true);
-      }
-    }
-    await this.queue.add(data._id, data, { delay, ...options });
-
-    return true;
-  }
-
-  private async addDelayJob(data: JobEntity, options: JobsOptions): Promise<boolean> {
-    const isDelayStep = data.type === StepTypeEnum.DELAY;
-
-    if (!isDelayStep) {
-      return false;
-    }
-
-    await this.jobRepository.updateStatus(data._id, JobStatusEnum.DELAYED);
-
-    const delay = WorkflowQueueService.calculateDelayAmount(data);
-
-    await this.queue.add(data._id, data, { delay, ...options });
-
-    return true;
-  }
-
-  private static checkValidDelayOverride(data: JobEntity): boolean {
-    if (!data.overrides?.delay) {
-      return false;
-    }
-    const values = Object.values(DigestUnitEnum);
-
-    return (
-      typeof data.overrides.delay.amount === 'number' &&
-      values.includes(data.overrides.delay.unit as unknown as DigestUnitEnum)
-    );
-  }
-
-  private static calculateDelayAmount(data: JobEntity): number {
-    if (data.step.metadata.type === DelayTypeEnum.SCHEDULED) {
-      const delayPath = data.step.metadata.delayPath;
-      const delayDate = data.payload[delayPath];
-      const delay = differenceInMilliseconds(new Date(delayDate), new Date());
-
-      if (delay < 0) {
-        throw new ApiException(`Delay date at path ${delayPath} must be a future date`);
-      }
-
-      return delay;
-    }
-
-    if (WorkflowQueueService.checkValidDelayOverride(data)) {
-      return WorkflowQueueService.toMilliseconds(
-        data.overrides.delay.amount as number,
-        data.overrides.delay.unit as DigestUnitEnum
-      );
-    }
-
-    return WorkflowQueueService.toMilliseconds(data.step.metadata.amount, data.step.metadata.unit);
-  }
-
-  private async delayedEventIsCanceled(job: JobEntity) {
-    if (job.type !== StepTypeEnum.DIGEST && job.type !== StepTypeEnum.DELAY) {
-      return false;
-    }
-    const count = await this.jobRepository.count({
-      _id: job._id,
-      status: JobStatusEnum.CANCELED,
-    });
-
-    return count > 0;
   }
 }
