+++ apps/api/src/app/events/usecases/add-job/add-job.command.ts
@@ -1,0 +1,10 @@
+import { IsBoolean, IsDefined } from 'class-validator';
+import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
+
+export class AddJobCommand extends EnvironmentWithUserCommand {
+  @IsDefined()
+  jobId: string;
+
+  @IsBoolean()
+  presend = false;
+}
