+++ apps/api/src/app/events/usecases/run-job/run-job.command.ts
@@ -1,0 +1,10 @@
+import { IsBoolean, IsDefined } from 'class-validator';
+import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
+
+export class RunJobCommand extends EnvironmentWithUserCommand {
+  @IsDefined()
+  jobId: string;
+
+  @IsBoolean()
+  presend = false;
+}
